Log out confirmation dialog (#6)

+++ forum-client/src/components/reusables/navigation-menu/navigation-menu.component.ts
@@ -1,7 +1,9 @@
 import { Component, inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
+import { MatDialog } from '@angular/material/dialog';
 import {MatToolbarModule} from '@angular/material/toolbar';
 import { Router } from '@angular/router';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'navigation-menu',
@@ -12,10 +14,21 @@ import { Router } from '@angular/router';
 export class NavigationMenuComponent {
 
   private readonly router = inject(Router)
+  private readonly dialog = inject(MatDialog)
 
   logOut(){
-    sessionStorage.clear()
-    this.router.navigate(["/login"])
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        dialogContent: "Do you want to log out?"
+      }
+    }).afterClosed().subscribe({
+      next: result => {
+        if(result == "Yes"){
+          sessionStorage.clear()
+          this.router.navigate(["/login"])
+        }
+      }
+    })
   }
 
   authenticated(): boolean {
